+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 
 ## [Unreleased]
 
+- Implemented SVG generation
+- Added property setting `is_paged` for Gate
 - Added check for already set `time_unit`
 - Made `create_directory_robust` remove then make
 - Minor code fixes